Safer account creation
I read few posts that people just copy/paste the passphrase to create their account but forget to store the passphrase and therefore can't do anything with the received ARK in that wallet, from a delegate or somebody else. I believe this is a better and safer approach of creating the account. This is also how Ledger Nano S works.